+++ Magic/src/main/java/com/elmakers/mine/bukkit/automata/AutomatonTemplate.java
@@ -2,6 +2,7 @@ package com.elmakers.mine.bukkit.automata;
 
 import java.util.Collection;
 import java.util.List;
+import java.util.logging.Logger;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -35,6 +36,10 @@ public class AutomatonTemplate {
 
     private final int playerRange;
     private final int minPlayers;
+    private final Integer minTimeOfDay;
+    private final Integer maxTimeOfDay;
+    private final Integer minPhaseOfMoon;
+    private final Integer maxPhaseOfMoon;
 
     public AutomatonTemplate(@Nonnull MageController controller, @Nonnull String key, @Nonnull ConfigurationSection configuration) {
         this.key = Preconditions.checkNotNull(key);
@@ -65,6 +70,64 @@ public class AutomatonTemplate {
         // Common parameters
         playerRange = configuration.getInt("player_range", 64);
         minPlayers = configuration.getInt("min_players", 0);
+        minTimeOfDay = parseTime(configuration, "min_time", controller.getLogger());
+        maxTimeOfDay = parseTime(configuration, "max_time", controller.getLogger());
+        if (configuration.contains("moon_phase")) {
+            minPhaseOfMoon = maxPhaseOfMoon = parseMoonPhase(configuration, "moon_phase", controller.getLogger());
+        } else {
+            minPhaseOfMoon = parseMoonPhase(configuration, "min_moon_phase", controller.getLogger());
+            maxPhaseOfMoon = parseMoonPhase(configuration, "max_moon_phase", controller.getLogger());
+        }
+    }
+
+    @Nullable
+    private Integer parseTime(ConfigurationSection configuration, String key, Logger log) {
+        Integer time = null;
+        if (configuration.contains(key)) {
+            if (configuration.isInt(key)) {
+                time = configuration.getInt(key);
+            } else {
+                String timeString = configuration.getString(key);
+                if (timeString.equalsIgnoreCase("day")) {
+                    time = 0;
+                } else if (timeString.equalsIgnoreCase("night")) {
+                    time = 13000;
+                } else if (timeString.equalsIgnoreCase("dusk") || timeString.equalsIgnoreCase("sunset")) {
+                    time = 12000;
+                } else if (timeString.equalsIgnoreCase("dawn") || timeString.equalsIgnoreCase("sunrise")) {
+                    time = 23000;
+                } else if (timeString.equalsIgnoreCase("noon") || timeString.equalsIgnoreCase("midday")) {
+                    time = 6000;
+                } else if (timeString.equalsIgnoreCase("midnight")) {
+                    time = 18000;
+                } else {
+                    log.warning("Invalid time in automata config: " + timeString);
+                }
+            }
+        }
+
+        return time;
+    }
+
+    @Nullable
+    private Integer parseMoonPhase(ConfigurationSection configuration, String key, Logger log) {
+        Integer phase = null;
+        if (configuration.contains(key)) {
+            if (configuration.isInt(key)) {
+                phase = configuration.getInt(key);
+            } else {
+                String phaseString = configuration.getString(key);
+                if (phaseString.equalsIgnoreCase("new")) {
+                    phase = 4;
+                } else if (phaseString.equalsIgnoreCase("full")) {
+                    phase = 0;
+                } else {
+                    log.warning("Invalid phase of moon in automata config: " + phaseString);
+                }
+            }
+        }
+
+        return phase;
     }
 
     @Nonnull
@@ -121,21 +184,49 @@ public class AutomatonTemplate {
         return effects;
     }
 
-    public boolean checkActive(Location location) {
-        if (minPlayers <= 0 || playerRange <= 0) {
-            return true;
+    private boolean checkRange(Integer min, Integer max, int value) {
+        if (min != null) {
+            if (max != null && max < min) {
+                if (value < min && value > max) {
+                    return false;
+                }
+            } else {
+                if (value < min) {
+                    return false;
+                }
+            }
+        }
+        if (max != null && value > max && (min == null || min <= max)) {
+            return false;
         }
 
-        int playerCount = 0;
-        int rangeSquared = playerRange * playerRange;
-        List<Player> players = location.getWorld().getPlayers();
-        for (Player player : players) {
-            if (player.getLocation().distanceSquared(location) <= rangeSquared) {
-                playerCount++;
+        return true;
+    }
+
+    public boolean checkActive(Location location) {
+        if (!checkRange(minTimeOfDay, maxTimeOfDay, (int)location.getWorld().getTime())) {
+            return false;
+        }
+        if (!checkRange(minPhaseOfMoon, maxPhaseOfMoon, (int)((location.getWorld().getFullTime() / 24000) % 8))) {
+            return false;
+        }
+
+        if (minPlayers >= 0 && playerRange > 0) {
+            int playerCount = 0;
+            int rangeSquared = playerRange * playerRange;
+            List<Player> players = location.getWorld().getPlayers();
+            for (Player player : players) {
+                if (player.getLocation().distanceSquared(location) <= rangeSquared) {
+                    playerCount++;
+                }
+            }
+
+            if (playerCount < minPlayers) {
+                return false;
             }
         }
 
-        return playerCount >= minPlayers;
+        return true;
     }
 
     public boolean isUndoAll() {