+++ Magic/CHANGELOG.md
@@ -6,7 +6,7 @@
  - Alter now works on flower pots (1.13 only)
  - Add some safety checks for class entity attributes getting stuck on server crash
  - Add jump_cast option for brooms and vehicles, to cast a spell with spacebar while riding
- - Add additional automata options, including nearby player requirements
+ - Add additional automata options, including nearby player, time of day and phase of moon requirements
 
 # 7.6.5